Move load_config/database to config.py
